requirements.txt
airtable==0.4.8
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
ems-dash-utils==0.0.1
finitelycomputable-flask-dispatcher==21.1
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.3
MarkupSafe==1.1.1
multitasking==0.0.9
ndexr-redditor==0.0.2
numpy==1.20.2
pandas==1.2.4
pdfkit==0.6.1
plotly==4.14.3
praw==7.2.0
prawcore==2.0.0
prometheus-client==0.10.1
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scikit-metrics==0.1.0
scipy==1.6.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
yfinance==0.1.59
